Avoid slow API calls for issue date
Issue archive page was loading really slowly since getting the
publication date of the latest post with that tag, for each tag, is
really costly. Instead, we use a script to set the tag description for
each issue tag and just use the description instead. This way, not only
is it more accurate, it is also faster. BUT it requires someone to set
the publication date of the issue on its tag. Given we have to upload
the front cover image anyway, it doesn't seem too much to also set the
date.

Signed-off-by: Timothy Langer <zeevox.dev@gmail.com>